requirements.txt
flask
aiohttp
discord.py-self
python-dotenv